Add flag to use AWSCLI or Bash for static upload
Some cases could result in a 403 Forbidden error when the included Bash
script is used to upload the static files archive to S3. This can be
avoided by using the AWS CLI on the same file system with the same set of
environment variables.

This commit adds a TF variable of boolean type to use the AWS CLI if set
to true. Users who come across the issue can set this flag (false by
default) to workaround the issue.

This commit also adds a few TF fmt changes.