Modify the output to be a map
The use case I was testing this needed to refer to a function version.
This was missing in the current TF config, so adding a publish=true
enables a version in the lambda function. The output is now a map of
ARN => version, which the calling modules can use to create additional
resources attached to the function.